requirements.txt
pymemcache
requests
PyYAML